% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_DM.R
\docType{data}
\name{DM_throat_info}
\alias{DM_throat_info}
\title{The sample template data for simulation - Dirichlet Multinomial Model}
\format{
List of 4.
}
\usage{
data(DM_throat_info)
}
\description{
The data before it's simulated with DM model in three scenarios:
common lineage, rare lineage, random OTUs.
}
\examples{
data(DM_throat_info)

}
\keyword{datasets}
